+++ src/main/java/core/basesyntax/FiguresAreaCalculator.java
@@ -1,5 +1,5 @@
 package core.basesyntax;
 
 public interface FiguresAreaCalculator {
-    int getArea();
+    double getArea();
 }
